fixed a bug in localstorage

+++ routes/ideas.js
@@ -45,7 +45,10 @@ router.post('/', async (req, res) => {
 //Update Idea
 router.put('/:id', async (req, res) => {
   try{
-    const updatedIdea = await Idea.findByIdAndUpdate(
+    const idea = await Idea.findById(req.params.id);
+    //Match the Usernames
+    if(idea.username === req.body.username){ 
+      const updatedIdea = await Idea.findByIdAndUpdate(
       req.params.id,
       {
         $set: {
@@ -55,8 +58,11 @@ router.put('/:id', async (req, res) => {
       },
       {new: true}
     );
-    res.json({success: true, data: updatedIdea});
+    return res.json({success: true, data: updatedIdea});
+    }
 
+    //Username does not match
+    res.status(403).json({success: false, error: 'You are not aurthorized to update this resource'});
   }catch(error){
     console.log(error);
     res.status(500).json({success: false, error: 'Something went wrong'});
@@ -66,8 +72,15 @@ router.put('/:id', async (req, res) => {
 //Delete Idea 
 router.delete('/:id', async (req, res) => {
   try{
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    //Match the Usernames
+    if(idea.username === req.body.username){
+      await Idea.findByIdAndDelete(req.params.id);
+    return res.json({success: true, data: {} });
+    }
+    // Username do not match
+    res.status(403).json({success: false, error: 'You are not aurthorized to delete this resource'});
   }catch(error){
     console.log(error);
     res.status(500).json({success: false, error: 'Something went Wrong'});
